+++ libraries/session/src/main/java/androidx/media3/session/MediaControllerImplLegacy.java
@@ -30,7 +30,6 @@ import android.media.AudioManager;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-import android.os.ResultReceiver;
 import android.os.SystemClock;
 import android.util.Pair;
 import android.view.Surface;
@@ -77,7 +76,6 @@ import androidx.media3.session.legacy.VolumeProviderCompat;
 import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
-import com.google.common.util.concurrent.SettableFuture;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -600,22 +598,13 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
 
   @Override
   public ListenableFuture<SessionResult> sendCustomCommand(SessionCommand command, Bundle args) {
-    if (controllerInfo.availableSessionCommands.contains(command)) {
+    if (controllerCompat != null) {
       controllerCompat.getTransportControls().sendCustomAction(command.customAction, args);
       return Futures.immediateFuture(new SessionResult(SessionResult.RESULT_SUCCESS));
+    } else {
+      return Futures.immediateFuture(
+          new SessionResult(SessionResult.RESULT_ERROR_SESSION_DISCONNECTED));
     }
-    SettableFuture<SessionResult> result = SettableFuture.create();
-    ResultReceiver cb =
-        new ResultReceiver(getInstance().applicationHandler) {
-          @Override
-          protected void onReceiveResult(int resultCode, Bundle resultData) {
-            result.set(
-                new SessionResult(
-                    resultCode, /* extras= */ resultData == null ? Bundle.EMPTY : resultData));
-          }
-        };
-    controllerCompat.sendCommand(command.customAction, args, cb);
-    return result;
   }
 
   @Override